#130: Refactored to use Bucket from bucketfs-java
