--- src/main/java/com/exasol/bucketfs/BucketConstants.java
@@ -1,14 +1,0 @@
-package com.exasol.bucketfs;
-
-/**
- * BucketFS constants.
- */
-public final class BucketConstants {
-    public static final String PATH_SEPARATOR = "/";
-    public static final String DEFAULT_BUCKETFS = "bfsdefault";
-    public static final String DEFAULT_BUCKET = "default";
-
-    private BucketConstants() {
-        // prevent instantiation
-    }
-}
